+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Pending
 
+- Clears launch arguments and launch environment values as part of tearDown, ready for the next test.
+- Added new `ActivitySheet` to represent the `UIActivityViewController` sheet.
 - Added support for `Carthage`, which allows project to use TABTestKit using `Carthage` as their dependency manager.
 
 ---